First draft of tksheet
I also commented out the widgets for now. For the tksheet grid, we want to have the list of entries (that is, the work number, status, and days until shipment) included after the first two rows of the spreadsheet because the first row contains the title and the search while the second row will have the section titles for the work number, status, and days left respectively. Let me know what you guys think about this plan I came up with.