Updated the Work In Progress
At this point, I decided to hold off implementing the MySQL commands until the upcoming 2nd sprint. Given that implementation of SQL queries can be a more difficult task, it is possible that I would run into an error that I am unable to connect to the database. It could also be due to the fact that every password in each .py file may be different. In the meantime, I added the quit button into the window so that it should give users the option to close out once they no longer need this window.